Cleans Empty Lines

+++ src/managers/cleansingprocess.py
@@ -1,6 +1,7 @@
 #import logfilemanager
-from .eventconfigmanager import EventConfig
-from .logfilemanager import LogFileManager
+print("NICE")
+from eventconfig import EventConfig
+from logfilemanager import LogFileManager
 import threading
 import os
 
@@ -8,20 +9,39 @@ class CleansingProcess:
     def __init__(self):
         #directory where cleansing will be needed 
         #self.directory
-        pass
+        self.eventConfig = EventConfig()
+        self.logfilemanager = LogFileManager()
+        self.remove_empty()
+        self.createLogFiles()
+        #self.fileList = []
 
     #Checks to see if file needs cleansing
-    def check_file(self):
-        pass
+    def createLogFiles(self):
+        print("IN GETFIles")
+        for dirName, subdirList, filelist in os.walk(self.eventConfig.getRootDir, topdown=False):
+            #print(dirName)
+            for fname in filelist:
+                #print("TESTING CREATION OF FILES")
+                #print(fname)
+                #print( dirName + "/" + fname)
+                #print(os.path.splitext(fname)[1])
+                self.logfilemanager.addLogFile(fname, dirName + "/" + fname, os.path.splitext(fname))
+        
 
     #Removes empty lines from a file
     #Files coming from event config
     def remove_empty(self):
-        with open(filePath) as in_file, open(filePath, 'r+') as out_file:
-            out_file.writelines(line for line in in_file if line.strip())
-            out_file.truncate()
+        print("IN REMOVE EMPTY")
+        for dirName, subdirList, filelist in os.walk(self.eventConfig.getRootDir, topdown=False):
+            for fname in filelist:
+                if (fname != '.DS_Store'):
+                    with open(dirName + "/" + fname) as in_file, open((dirName + "/" + fname), 'r+') as out_file:
+                        print(fname)
+                        out_file.writelines(line for line in in_file if line.strip())
+                        out_file.truncate()
+                
 
+    #TODO
     #removes unwanted characters from a file
     def remove_characters(self):
         pass
-
